+++ packages/common/constants/index.js
@@ -1,5 +1,5 @@
 export * from './lib/access';
-export * from './lib/appEvent';
+export * from './lib/events';
 export * from './lib/assets';
 export * from './lib/attributes';
 export * from './lib/auth';