+++ packages/common/constants/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.377.0](https://github.com/DEIPworld/deip-modules/compare/v1.376.0...v1.377.0) (2022-05-13)
+
+
+### Features
+
+* **@deip/constants:** add chainDomain events ([d82878f](https://github.com/DEIPworld/deip-modules/commit/d82878fb9ad5af632ac25cd4efd01021a9a00605))
+* **@deip/reviews-module:** add reviews module ([7cb8c88](https://github.com/DEIPworld/deip-modules/commit/7cb8c889114f79faa2810bf02326fac88c7b37ec))
+
+
+
+
+
 # [1.376.0](https://github.com/DEIPworld/deip-modules/compare/v1.375.1...v1.376.0) (2022-04-08)